requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
firebase-admin
paramiko
python-dotenv
httpx